+++ docs/changelog.rst
@@ -11,7 +11,9 @@ License
 -----
 :Released: under development
 
-* Improvement: Adds :ref:`needs_build_json` config option to build ``needs.json`` in parallel to other output formats
+* Improvement: :ref:`filter_func` support arguments.
+  (`#429 <https://github.com/useblocks/sphinxcontrib-needs/issues/429>`_)
+* Improvement: Adds :ref:`needs_build_json` config option to build ``needs.json`` in parallel to other output formats.
   (`#485 <https://github.com/useblocks/sphinxcontrib-needs/issues/485>`_)